+++ includes/Base/AdminNotice.php
@@ -43,7 +43,7 @@ class AdminNotice extends BaseController
 				unset( $_GET['activate'] ); // phpcs:ignore
 			}
 			/* translators: %s: Plugin Name */
-			echo '<div class="error"><p>' . sprintf( __( 'Your version of WordPress is below the minimum version of WordPress required by %s plugin. Please upgrade WordPress to 5.2.0 or later.', 'migrate-wp-cron-to-action-scheduler' ), $this->name ) . '</p></div>';
+			echo '<div class="error"><p>' . sprintf( esc_html__( 'Your version of WordPress is below the minimum version of WordPress required by %s plugin. Please upgrade WordPress to 5.2.0 or later.', 'migrate-wp-cron-to-action-scheduler' ), esc_html( $this->name ) ) . '</p></div>';
 		    return;
 		}
 		
@@ -54,7 +54,7 @@ class AdminNotice extends BaseController
 				unset( $_GET['activate'] ); // phpcs:ignore
 			}
 			/* translators: %s: Plugin Name */
-			echo '<div class="error"><p>' . sprintf( __( 'Your version of PHP is below the minimum version of PHP required by %s plugin. Please contact your host and request that your version be upgraded to 5.6 or later.', 'migrate-wp-cron-to-action-scheduler' ), $this->name ) . '</p></div>';
+			echo '<div class="error"><p>' . sprintf( esc_html__( 'Your version of PHP is below the minimum version of PHP required by %s plugin. Please contact your host and request that your version be upgraded to 5.6 or later.', 'migrate-wp-cron-to-action-scheduler' ), esc_html( $this->name ) ) . '</p></div>';
 			return;
 		}
 
@@ -63,7 +63,7 @@ class AdminNotice extends BaseController
 			<div class="notice notice-success">
 				<p><strong><?php
 				/* translators: %s: Plugin Name */ 
-				printf( __( 'Thanks for installing %1$s v%2$s plugin. Click <a href="%3$s">here</a> to view Action Scheduler tasks.', 'migrate-wp-cron-to-action-scheduler' ), $this->name, $this->version, admin_url( 'tools.php?page=action-scheduler' ) ); ?></strong></p>
+				printf( esc_html__( 'Thanks for installing %1$s v%2$s plugin. Click <a href="%3$s">here</a> to view Action Scheduler tasks.', 'migrate-wp-cron-to-action-scheduler' ), 'Advanced Cron Scheduler', esc_html( ACSWP_PLUGIN_VERSION ), esc_url( admin_url( 'tools.php?page=action-scheduler' ) ) ); ?></strong></p>
 			</div> <?php
 		    delete_transient( 'acswp-show-notice-on-activation' );
 		}
@@ -78,7 +78,7 @@ class AdminNotice extends BaseController
 		if ( $show_rating ) {
 			$dismiss = wp_nonce_url( add_query_arg( 'acswp_notice_action', 'dismiss_rating' ), 'acswp_notice_nonce' );
 			$no_thanks = wp_nonce_url( add_query_arg( 'acswp_notice_action', 'no_thanks_rating' ), 'acswp_notice_nonce' ); ?>
-			
+
 			<div class="notice notice-success">
 				<p><?php echo wp_kses_post( 'Hey, I noticed you\'ve been using Advanced Cron Scheduler for more than 1 week – that’s awesome! Could you please do me a BIG favor and give it a <strong>5-star</strong> rating on WordPress? Just to help us spread the word and boost my motivation.', 'migrate-wp-cron-to-action-scheduler' ); ?></p>
             	<p><a href="https://wordpress.org/support/plugin/migrate-wp-cron-to-action-scheduler/reviews/?filter=5#new-post" target="_blank" class="button button-secondary" rel="noopener"><?php esc_html_e( 'Ok, you deserve it', 'migrate-wp-cron-to-action-scheduler' ); ?></a>&nbsp;
@@ -98,7 +98,7 @@ class AdminNotice extends BaseController
 		if ( $show_donate ) {
 			$dismiss = wp_nonce_url( add_query_arg( 'acswp_notice_action', 'dismiss_donate' ), 'acswp_notice_nonce' );
 			$no_thanks = wp_nonce_url( add_query_arg( 'acswp_notice_action', 'no_thanks_donate' ), 'acswp_notice_nonce' ); ?>
-			
+
 			<div class="notice notice-success">
 				<p><?php echo wp_kses_post( 'Hey, I noticed you\'ve been using Advanced Cron Scheduler for more than 2 week – that’s awesome! If you like Advanced Cron Scheduler and you are satisfied with the plugin, isn’t that worth a coffee or two? Please consider donating. Donations help me to continue support and development of this free plugin! Thank you very much!', 'migrate-wp-cron-to-action-scheduler' ); ?></p>
 				<p><a href="https://www.paypal.me/iamsayan" target="_blank" class="button button-secondary" rel="noopener"><?php esc_html_e( 'Donate Now', 'migrate-wp-cron-to-action-scheduler' ); ?></a>&nbsp;
@@ -133,7 +133,7 @@ class AdminNotice extends BaseController
 
 		check_admin_referer( 'acswp_notice_nonce' );
 
-		$notice = sanitize_text_field( $_REQUEST['acswp_notice_action'] );
+		$notice = sanitize_text_field( wp_unslash( $_REQUEST['acswp_notice_action'] ) );
 		$notice = explode( '_', $notice );
 		$notice_type = end( $notice );
 		array_pop( $notice );
@@ -153,7 +153,7 @@ class AdminNotice extends BaseController
 			}
 		}
 	
-		wp_redirect( remove_query_arg( [ 'acswp_notice_action', '_wpnonce' ] ) );
+		wp_safe_redirect( remove_query_arg( [ 'acswp_notice_action', '_wpnonce' ] ) );
 		exit;
 	}